move validateColumnSelection to tablatureEditorStore dir
