+++ src/modules/tablatureEditorStore/utils/validateColumnSelection.ts
